+++ frontend/src/pages/recipes/RecipesPage.tsx
@@ -3,7 +3,7 @@ import {useEffect} from "react";
 type RecipesScreenProps = {
     setCurrentRoute : (url:string) => void
 }
-export default function RecipesScreen(props: Readonly<RecipesScreenProps>) {
+export default function RecipesPage(props: Readonly<RecipesScreenProps>) {
 
     const url = window.location.href;
