+++ frontend/src/pages/registration/RegistrationPagen.css
