+++ frontend/src/pages/start/DesktopStartPage.css
