+++ frontend/src/pages/start/StartPage.css
@@ -15,7 +15,7 @@
 }
 
 .image-hero{
-    background-image: url("../assets/startpage-imgs/img-slider-1.jpg");
+    background-image: url("../../assets/startpage-imgs/img-slider-1.jpg");
     background-size: cover;
     background-repeat: no-repeat;
     border-radius: 16px;
